Document new license, small refinements, reformat for new developer
The original author was contacted and was okay as long as there was some
attribution.  Since it is handy, GPLv3 has been chosen.  Add the magic
number for scripts to the top of files, no need to explicitly call Python
to run them on a POSIX system.  Tabs are better than spaces.