Implement extraction as full image files
This version now has the capability to extract the DZ file as images of
entire slices^Wpartitions, or entire disk/flash image.  This version will
also interpret extra arguments as chunk or slice numbers to extract,
rather than needing a separate command flag for indicating whether to
extract everything versus just one.  Several of the previously unknown
fields in the header are identified and verification of this has been
implemented, but one last field is still currently unknown.