Add notes for the fields in chunk headers, pad out files as headers indicate
More documentation of what I believe the chunk header fields mean.  Of
note, what had previously been an unknown field appears to be a count of
blocks that should be wiped before writing data.  The code now creates
holes to match these areas, and as such the system.img file can be
directly mounted (previously it had problems).  This seems a worrisome
approach with the "cust" slice/filesystem (an area of the /data
filesystem is *erased* during upgrades!).