Improve handling of slices, centralize certain values
We now detect if there are unallocated gaps between slices and create
markers for them.  This could be of high value if someone was to stash
magic data in these areas.  Move the shift to block size value and file
handle onto the DZFile, this should make things if it turns out we need
to probe to figure out what the block size is.