Add saving of original file chunks during chunk extraction
I'm pretty sure I've got the chunk headers pretty well completely figured
out, but there is a bit of room for interpretation.  Notably, does the
wipe count indicate blocks that will be wiped to zero, blocks that will
be trimmed?  In particular the "cust" area where only some portions get
wiped is suspicious.  As such I'd prefer to leave some portions
untouched, so add support for stashing the original chunks.  As a
simplification, do a chdir() to the output directory, which avoids all
the calls to os.path.join().