Add .params file for rebuilding slices feature
Appears certain "magic" values will be invaluable when recreating DZ
files, so save the valuable ones with extracted slices.  The block wipe
count is of particular note (looks worrisome on the "cust" and
"BackupGPT" slices).
Add a -c/--chunks option, may be a better name than -x.  Explicitly close
files being written to.  Seems this hasn't been a problem so far, but on
the principle of being safe...