Initial implementation of image2chunks.py
This is a tool written to convert images into chunk files.  The chunk
files would then in theory be merged into a DZ file.  Right now this
utterly fails to generate separate chunk files, and merely produces one
giant chunk file, but this IS a work in progress.  I'm suspecting this
may in fact be a weakness of the FS I'm storing data on, perhaps this
filesystem doesn't expose holes with SEEK_HOLE/SEEK_DATA.