Misc refinements, do pack()/unpack() in parent class, message consistency
The operation of packing and unpacking the header structures is likely to
be used often, move some of this onto DZStruct.  Also handles some of the
common operations of such actions.

Simplify the targetAddr comparison in DZChunk's initialization (compare
as block#, therefore no shifting).
Fix the bug of assuming .params files are always going to need lastWipe,
the ones not containing any chunks won't need this data.

argparse seems to be designed for help messages starting with lowercase
(notice the message for -h).  Lowercase the messages for various options.