Initial implementation of mkdz.py
This script builds what appears to be a valid DZ file from the chunks and
saved header in a directory.  Since the goal is merely to build a valid
DZ file, not as much verification is done as in undz.py.  Currently only
two fields in the header are known how to be computed, so the rest are
copied from a known valid file.  Hopefully these two together are
sufficient to produce a DZ file considered valid by LGE's tools.