Move internal utilities to a subdir
Currently there are only two utility files, but the base directory
already has a number of files in it, so moving the utilities may be high
value for the future.  Needed a bit of reading to figure out Python's
mechanism for this case, but get this done now rather than later.