Fix end of headers handling
Appears this format doesn't work the way IOMonster/thecubed thought.
There is a "\x03" (control-C) *between* the headers which indicates the
next header is last.  This also means the last one is misaligned.  Now
we get the actual data.  Check the area between the headers and data
start for non-zeros, which would be a warning sign.

Add os.SEEK_xxx to all the seeks.  SEEK_CUR is the default, but make that
more obvious.