Modify parameter file names (beware modification needs update!)
The dotfiles being used to save information about KDZ and DZ files are
being renamed.  The KDZ information was previously stored in ".params"
and that is now ".kdz.params".  The DZ information was previously stored
in ".header", but has now been moved to ".dz.header".  If you were in the
middle of a project, you WILL need to rename these files!

The goal of this change is to avoid confusing the filenames.  I hope to
move to using a .params type text-file for storing information about the
DZ header at which point we might need two files named ".params" if
everything was in a single directory.  This fixes that issue.

Teeny fix to undz.py, note that the values are not used by `mkdz` (also
not used by `mkkdz`, but that seems obvious).