Move to loading values out of .dz.params, make mkdz.py fully functional again
Had planned to have mkdz.py load values out of .dz.params for a while,
but due to the necessity of having a blockShift value this has finally
happened.  DZ files for newer devices use a non-512 byte block size, need
to deal with that.  Sorting of chunks is now a bit funkier.  We now
successfully handle DZ files for LG's G5 devices.