Bugfix on mkdz.py, .dz.params comment fix
Originally when undz.py started creating the .dz.params files they were
strictly commentary and didn't do anything.  Now mkdz.py uses the files
so the comment about them being reference-only is no longer valid.  The
old date code in DZ files is a string, not an integer.  Alas the value
matches as a string so it needs special care.