Ensure we don't try to truncate files to negative length
More recent devices are getting chunks which fall outside the bounds of
what had been previously observed.  Sometimes these are causing some
interesting code paths to be touched.  The negative length indicates we
don't know what the parameters of the slice are and slices like that are
showing up.  Since these are causing problems for some users of undz, we
have to work around them for now.