Initial handling of extra data
Another sanity test triggered indicating the presence of unknown data.
An initial look seems to indicate this data includes a short index of the
DZ data, and could include a signature (ends any hope of creating
modified files).  This also means unkdz.py and undz.py will need to be
merged in order to verify this index.